added auth, protect routes

+++ routes/api.php
@@ -12,10 +12,16 @@ Route::get('/user', function (Request $request) {
 Route::prefix('auth')->group(function () {
     Route::post('/register', [AuthController::class, 'registerUser']);
     Route::post('/login', [AuthController::class, 'loginUser']);
-    Route::get('/users', [AuthController::class, 'listUsers']);
-    Route::get('/users/{id}', [AuthController::class, 'userProfile']);
-    Route::delete('/users/{id}', [AuthController::class, 'deleteUser']);
     Route::post('/logout', [AuthController::class, 'logoutUser']);
 });
 
-Route::apiResource('students', StudentController::class);
+Route::apiResource('students', StudentController::class)->middleware('auth:sanctum');
+Route::get('/users', [AuthController::class, 'listUsers'])->middleware('auth:sanctum');
+Route::get('/users/{id}', [AuthController::class, 'userProfile'])->middleware('auth:sanctum');
+Route::delete('/users/{id}', [AuthController::class, 'deleteUser'])->middleware('auth:sanctum');
+
+Route::post('/tokens/create', function (Request $request) {
+    $token = $request->user()->createToken($request->token_name);
+
+    return ['token' => $token->plainTextToken];
+});
